+++ ruma-signatures/CHANGELOG.md
@@ -5,3 +5,7 @@ Breaking changes:
 * Remove `Copy` implementation for `Algorithm`
 * Remove `Copy` and `Clone` implementations for `Ed25519Verifier`
 * Upgrade ruma-identifiers
+
+Bug fixes:
+
+* Verify only the required signatures on `verify_event`